WIP - represent a set of re-writes using a const list
this is needed because each re-write has its own `Rewrite::Ctx` type, so
it can't be converted into a `dyn Rewrite` trait object without
specifying the type of `Ctx` (since we ned to know its size at compile
time).

we could use `Box<dyn Any>` or something like that for `Ctx` but that
would have a great runtime overhead.