there is a bug in the code...
for some reason, loops are created in the egraph, even though my code
should not allow them.

no idea what the fuck is going on.

the only thing i know is that if you remove the call to
`UnionFind::set_parent_of_item` inside of `UnionFind::root_of_item`,
the bug stops happeneing. what the fuck man?

figure it out and fix it...